Set React Context and useContext
